# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A comprehensive Point of Sale (POS) backend system built with Go and Fiber framework. This API provides endpoints for managing users, businesses, outlets, products, sales, customers, and all other POS-related operations.",
        "title": "POS Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/indrayyana/go-fiber-boilerplate/blob/main/LICENSE"
        },
        "version": "1.0.0"
    },
    "host": "localhost:3000",
    "basePath": "/v1",
    "paths": {
...
        "/sales": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all sales with pagination and search functionality",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sale"
                ],
                "summary": "Get all sales",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Maximum number of sales",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by invoice number or customer name",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by outlet ID",
                        "name": "outletId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by customer ID",
                        "name": "customerId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status (pending, completed, cancelled)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by date from (YYYY-MM-DD)",
                        "name": "dateFrom",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by date to (YYYY-MM-DD)",
                        "name": "dateTo",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessWithPaginatedSales"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDetails"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDetails"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDetails"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new sale transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sale"
                ],
                "summary": "Create new sale",
                "parameters": [
                    {
                        "description": "Sale data",
                        "name": "sale",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validation.CreateSale"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessWithSale"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDetails"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDetails"
                        }
                    },
                    "409": {
                        "description": "Invoice number already exists",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDetails"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDetails"
                        }
                    }
                }
            }
        },
        "/sales/invoice/{invoiceNumber}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a specific sale by its invoice number",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sale"
                ],
                "summary": "Get sale by invoice number",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invoice Number",
                        "name": "invoiceNumber",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessWithSale"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDetails"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDetails"
                        }
                    },
                    "404": {
                        "description": "Sale not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDetails"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDetails"
                        }
                    }
                }
            }
        },
        "/sales/report": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get sales report with various aggregations",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sale"
                ],
                "summary": "Get sales report",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by outlet ID",
                        "name": "outletId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by date from (YYYY-MM-DD)",
                        "name": "dateFrom",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by date to (YYYY-MM-DD)",
                        "name": "dateTo",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Group by (day, week, month, year)",
                        "name": "groupBy",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessWithSalesReport"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDetails"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDetails"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDetails"
                        }
                    }
                }
            }
        },
        "/sales/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a specific sale by its ID with items details",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sale"
                ],
                "summary": "Get sale by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sale ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessWithSale"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDetails"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDetails"
                        }
                    },
                    "404": {
                        "description": "Sale not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDetails"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDetails"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing sale",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sale"
                ],
                "summary": "Update sale",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sale ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Sale data to update",
                        "name": "sale",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validation.UpdateSale"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessWithSale"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDetails"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDetails"
                        }
                    },
                    "404": {
                        "description": "Sale not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDetails"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDetails"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a sale by ID (soft delete)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sale"
                ],
                "summary": "Delete sale",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sale ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Common"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDetails"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDetails"
                        }
                    },
                    "404": {
                        "description": "Sale not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDetails"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDetails"
                        }
                    }
                }
            }
        },
        "/sales/{id}/status": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update the status of a sale (pending, completed, cancelled)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sale"
                ],
                "summary": "Update sale status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sale ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Sale status data",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validation.UpdateSaleStatus"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessWithSale"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDetails"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDetails"
                        }
                    },
                    "404": {
                        "description": "Sale not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDetails"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDetails"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.ErrorDetails": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "errors": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "response.SuccessWithPaginatedSales": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "limit": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Sale"
                    }
                },
                "status": {
                    "type": "string"
                },
                "total_pages": {
                    "type": "integer"
                },
                "total_results": {
                    "type": "integer"
                }
            }
        },
...
        "validation.CreateSale": {
            "type": "object",
            "required": [
                "grand_total",
                "invoice_number",
                "outlet_id",
                "outlet_staff_id",
                "status",
                "table_id",
                "total"
            ],
            "properties": {
                "customer_id": {
                    "type": "string"
                },
                "discount": {
                    "type": "number",
                    "minimum": 0
                },
                "grand_total": {
                    "type": "number",
                    "minimum": 0
                },
                "invoice_number": {
                    "type": "string",
                    "maxLength": 50
                },
                "note": {
                    "type": "string"
                },
                "outlet_id": {
                    "type": "string"
                },
                "outlet_staff_id": {
                    "type": "string"
                },
                "payment_method_id": {
                    "type": "string"
                },
                "sale_date": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "paid",
                        "unpaid",
                        "void",
                        "hold"
                    ]
                },
                "table_id": {
                    "type": "string"
                },
                "tax": {
                    "type": "number",
                    "minimum": 0
                },
                "total": {
                    "type": "number",
                    "minimum": 0
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Example Value: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}